better smoothing algorithm
